[WIP] Cluster restart lifecycle operation for k8s
WIP: Bake container agent image into Atomic image
  - Image to test with: https://test-fedora-23-ironic-image.web.cern.ch/test-fedora-23-ironic-image/fedora-atomic-23-drago.qcow2

This patch adds support for cluster lifecycle operations, starting with
the restart operation. The restart action restarts the COE services on
the master and worker nodes. The endpoint does not accept a request
body and responds with a 204 with no response body. The endpoint is:

PATCH /clusters/{id}/actions/restart

Cluster lifecycle operations that deal with software on a node (such as
restart) use a containerized version of the software config agent
(os-collect-config and friends) that can poll Heat to receive
software updates.

Change-Id: I4f5a9f01726fa28dde6a1be7bdce5f79a31e8045
Depends-On: I2ff7fdcd75d9700c0584a12c3f01a509e4b5725e
Depends-On: Ia49a549e11e2e6a1e9399a1370f990780c17bf0d
Depends-on: I9f44803cf04b572074bdb4b97ee58827c5611514
Partially-Implements: blueprint cluster-restart-operation